athenaos_backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
vaderSentiment>=3.3.2
python-multipart>=0.0.9
google-genai>=0.7.0
sentence-transformers>=2.6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
httpx>=0.26.0
# Optional — for video transcription:
openai-whisper>=20231117
